\name{perryExamples-package}
\alias{perryExamples-package}
\alias{perryExamples}
\docType{package}
\title{
Examples for integrating prediction error estimation into
regression models
}
\description{
Examples for integrating package perry for prediction error estimation into regression models.
}
\details{
\tabular{ll}{
Package: \tab perryExamples\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2014-02-21\cr
Depends: \tab R (>= 2.14.1), perry, robustbase\cr
Imports: \tab stats, quantreg, lars\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Index:
\preformatted{
Bundesliga              Austrian Bundesliga football player data
TopGearMPG              Top Gear fuel consumption data
ladlasso                LAD-lasso with penalty parameter selection
lasso                   Lasso with penalty parameter selection
perry-methods           Resampling-based prediction error for fitted
                        models
perryExamples-package   Examples for integrating prediction error
                        estimation into regression models
ridge                   Ridge regression with penalty parameter
                        selection
}
}
\author{
Andreas Alfons [aut, cre]

Maintainer: Andreas Alfons <alfons@ese.eur.nl>
}
\keyword{package}
