\name{perryExamples-package}
\alias{perryExamples-package}
\alias{perryExamples}
\docType{package}
\title{
\packageTitle{perryExamples}
}
\description{
\packageDescription{perryExamples}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{perryExamples}
\packageIndices{perryExamples}
}
\author{
\packageAuthor{perryExamples}

Maintainer: \packageMaintainer{perryExamples}
}
\keyword{package}
